added panda
This script scrapes MarketWatch for AMD stock Prices. It will write it to several csv files, amd.csv has the financial numbers, amdprice.csv has current, open, day high, day low, and day average. 
Panda takes the amdprice.csv and rewrites a new file (output.csv) that is only concerned with the Day Average column, it strips the string Day Average and only writes the prices. 
Then it will get the mean of all of the prices in the Day Average Column.